SDformer_hw_32517499/requirements.txt
numpy
pandas
PyYAML
tqdm
matplotlib
opencv-python
imageio
numba
h5py
tables
hdf5plugin
mlflow
einops
timm==0.6.13
spikingjelly==0.0.0.0.14
omegaconf

